requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.9.2
accelerate=1.3.0=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
asgiref=3.8.1=pypi_0
asttokens=3.0.0=pypi_0
async-timeout=4.0.3=pypi_0
attrs=25.1.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.2.1=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py310h5da7b33_9
build=1.2.2.post1=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.12.31=haa95532_0
cachetools=5.5.1=pypi_0
certifi=2024.12.14=py310haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.6.3=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pypi_0
contourpy=1.3.1=pypi_0
cuda-cccl=12.8.55=0
cuda-cccl_win-64=12.8.55=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.8.55=0
cuda-opencl-dev=12.8.55=0
cuda-profiler-api=12.8.55=0
cuda-runtime=12.1.0=0
cuda-version=12.8=3
cycler=0.12.1=pypi_0
dataclasses-json=0.6.7=pypi_0
datasets=3.2.0=pypi_0
debugpy=1.8.12=pypi_0
decorator=5.1.1=pypi_0
deprecated=1.2.18=pypi_0
dill=0.3.8=pypi_0
durationpy=0.9=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.2.0=pypi_0
fastapi=0.115.8=pypi_0
filelock=3.13.1=py310haa95532_0
flatbuffers=25.1.24=pypi_0
fonttools=4.55.8=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.5.0=pypi_0
fsspec=2024.9.0=pypi_0
giflib=5.2.2=h7edc060_0
gmp=6.3.0=h537511b_0
gmpy2=2.2.1=py310h827c3e9_0
google-auth=2.38.0=pypi_0
googleapis-common-protos=1.66.0=pypi_0
greenlet=3.1.1=pypi_0
grpcio=1.67.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
httpx-sse=0.4.0=pypi_0
huggingface-hub=0.28.1=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=py310haa95532_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.5.2=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.5=pypi_0
ipython=8.32.0=pypi_0
ipywidgets=8.1.5=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.5=py310haa95532_0
joblib=1.4.2=pypi_0
jpeg=9e=h827c3e9_3
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
khronos-opencl-icd-loader=2024.05.08=h8cc25b3_0
kiwisolver=1.4.8=pypi_0
kubernetes=32.0.0=pypi_0
langchain=0.3.17=pypi_0
langchain-chroma=0.2.1=pypi_0
langchain-community=0.3.16=pypi_0
langchain-core=0.3.33=pypi_0
langchain-milvus=0.1.8=pypi_0
langchain-text-splitters=0.3.5=pypi_0
langsmith=0.3.6=pypi_0
lcms2=2.16=hb4a4139_0
lerc=4.0.0=h5da7b33_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.9.55=0
libcurand-dev=10.3.9.55=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.22=h5bf469e_0
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=h44ae7cf_1
libuv=1.48.0=h827c3e9_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h3d04722_1
lz4-c=1.9.4=h2bbff1b_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py310h827c3e9_1
marshmallow=3.26.1=pypi_0
matplotlib=3.10.0=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py310h827c3e9_2
mkl_fft=1.3.11=py310h827c3e9_0
mkl_random=1.2.8=py310hc64d2fc_0
mmh3=5.1.0=pypi_0
monotonic=1.6=pypi_0
mpc=1.3.1=h827c3e9_0
mpfr=4.2.1=h56c3642_0
mpmath=1.3.0=py310haa95532_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
mypy-extensions=1.0.0=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.4.2=py310haa95532_0
nltk=3.9.1=pypi_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.20.1=pypi_0
openjpeg=2.5.2=hae555c5_0
openssl=3.0.15=h827c3e9_0
opentelemetry-api=1.30.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.30.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.30.0=pypi_0
opentelemetry-instrumentation=0.51b0=pypi_0
opentelemetry-instrumentation-asgi=0.51b0=pypi_0
opentelemetry-instrumentation-fastapi=0.51b0=pypi_0
opentelemetry-proto=1.30.0=pypi_0
opentelemetry-sdk=1.30.0=pypi_0
opentelemetry-semantic-conventions=0.51b0=pypi_0
opentelemetry-util-http=0.51b0=pypi_0
orjson=3.10.15=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pypi_0
pillow=11.1.0=py310h096bfcc_0
pip=25.0=py310haa95532_0
platformdirs=4.3.6=pypi_0
posthog=3.11.0=pypi_0
prompt-toolkit=3.0.50=pypi_0
propcache=0.2.1=pypi_0
protobuf=5.29.3=pypi_0
psutil=6.1.1=pypi_0
pure-eval=0.2.3=pypi_0
pyarrow=19.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydantic-settings=2.7.1=pypi_0
pygments=2.19.1=pypi_0
pymilvus=2.5.4=pypi_0
pyparsing=3.2.1=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
pyreadline3=3.5.4=pypi_0
pysocks=1.7.1=py310haa95532_0
python=3.10.16=h4607a30_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytorch=2.5.1=py3.10_cuda12.1_cudnn9_0
pytorch-cuda=12.1=hde6ce7c_6
pytorch-mutex=1.0=cuda
pytz=2025.1=pypi_0
pywin32=308=pypi_0
pyyaml=6.0.2=py310h827c3e9_0
pyzmq=26.2.1=pypi_0
rank-bm25=0.2.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=py310haa95532_1
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.9.4=pypi_0
rsa=4.9=pypi_0
safetensors=0.5.2=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.1=pypi_0
sentence-transformers=3.4.1=pypi_0
setuptools=75.8.0=py310haa95532_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.37=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
starlette=0.45.3=pypi_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=h59b6b97_0
tenacity=9.0.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h0416ee5_0
tokenizers=0.21.0=pypi_0
tomli=2.2.1=pypi_0
torchaudio=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.48.2=pypi_0
typer=0.15.1=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.12.2=py310haa95532_0
tzdata=2025.1=pypi_0
ujson=5.10.0=pypi_0
urllib3=2.3.0=py310haa95532_0
uvicorn=0.34.0=pypi_0
vc=14.40=haa95532_2
vs2015_runtime=14.42.34433=h9531ae6_2
watchfiles=1.0.4=pypi_0
wcwidth=0.2.13=pypi_0
websocket-client=1.8.0=pypi_0
websockets=14.2=pypi_0
wheel=0.45.1=py310haa95532_0
widgetsnbextension=4.0.13=pypi_0
win_inet_pton=1.1.0=py310haa95532_0
wrapt=1.17.2=pypi_0
xxhash=3.5.0=pypi_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.18.3=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h8cc25b3_1
zstandard=0.23.0=pypi_0
zstd=1.5.6=h8880b57_0
